requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
aniso8601==9.0.1
async-timeout==4.0.2
attrs==22.1.0
black==22.6.0
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
Deprecated==1.2.13
flake8==5.0.4
flasgger==0.9.5
Flask==2.2.1
Flask-JWT-Extended==4.4.3
Flask-Limiter==2.6.0
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
frozenlist==1.3.1
gevent==21.12.0
greenlet==1.1.2
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.9.1
limits==2.7.0
MarkupSafe==2.1.1
marshmallow==3.17.1
marshmallow-sqlalchemy==0.28.1
mccabe==0.7.0
mistune==2.0.4
multidict==6.0.2
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.9.1
pydantic==1.9.2
pyflakes==2.5.0
Pygments==2.12.0
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.2
pytest-asyncio==0.19.0
python-dotenv==0.20.0
pytz==2022.2
PyYAML==6.0
redis==4.3.4
rich==12.5.1
six==1.16.0
SQLAlchemy==1.4.40
tomli==2.0.1
typing_extensions==4.3.0
Werkzeug==2.2.1
wrapt==1.14.1
yarl==1.8.1
zope.event==4.5.0
zope.interface==5.4.0
